refactor: address review feedback — flatten modules, split tests, fix types
Addresses all review feedback from @mkmeral and automated review:

1. **Flatten module structure** (per @mkmeral): Remove sub-packages
   (shell/, editor/, python_repl/ directories). Tools are now flat
   files under src/strands/vended_tools/ — aligns with 'Prefer Flat
   Namespaces' decision record.

2. **Split test file** (per @mkmeral): The single 1161-line
   test_vended_tools.py is now split into test_shell.py, test_editor.py,
   test_python_repl.py, and test_init.py — mirrors src/ structure per
   AGENTS.md convention.

3. **Fix deprecated asyncio pattern**: Replace asyncio.get_event_loop()
   with @pytest.mark.asyncio + async def tests throughout.

4. **Type sandbox as Sandbox**: Editor helper functions now use proper
   Sandbox type instead of Any — enables mypy checking and IDE support.

5. **Use read_text()/write_text()**: Replace manual .decode()/.encode()
   with the Sandbox convenience API as designed.

6. **Optimize cwd tracking**: Append '; echo __STRANDS_CWD__; pwd' to
   commands instead of running a separate pwd call after every execution
   — halves sandbox interactions for remote sandboxes.

7. **Extract shared _get_config**: New _utils.py with get_tool_config()
   eliminates the copy-pasted helper across all three tools.

8. **Narrow except clause**: Replace broad 'except Exception' catch-all
   with targeted (FileNotFoundError, UnicodeDecodeError, OSError,
   ValueError) — unexpected errors now propagate for debugging.

All 65 tests pass.
